// src/map/types.ts
export enum MapScope {
    Summary = 'summary',
    Mine = 'mine',
    Around = 'around',
    Cells = 'cells',
    All = 'all',
}
// ...
export interface AroundQuery {
    x: number;
    y: number;
    radius: number;
}

export interface MapQuery {
    scope: MapScope;
    tokenIds: Array<string> | null;
    around: AroundQuery | null;
    ownerAddress: string | null;
}